--- a/velox/functions/lib/RowsTranslationUtil.h
+++ b/velox/functions/lib/RowsTranslationUtil.h
@@ -13,6 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
+#pragma once
+
 #include "velox/common/base/Nulls.h"
 #include "velox/vector/BaseVector.h"
 #include "velox/vector/SelectivityVector.h"